Add market status/result/open_time/liquidity fields, resolved sweep, and FreshMarketStrategy
- Migrations 0012/0013: add open_time, status, result, last_price, liquidity to markets table
- Propagate new fields through KalshiMarketSchema, MarketRow, Market dataclass, repository upsert, and all _market_row_to_domain() sites
- MarketWatcher._sweep_closed_markets(): re-fetches markets past close_time or early-closed with open positions so they get marked resolved/finalized
- PositionMonitor: prefer market.result ("yes"/"no") for resolution inference; read market.status directly instead of metadata["status"]
- days_to_close: switch from integer .days to fractional float (total_seconds/86400); type StrategyConfig fields as float
- FreshMarketStrategy: targets markets opened within 24h, spread ≤ 0.15, max 7-day horizon
- SPEC.md Phase 3 table reformatted with issue links; adds T47/T48 (OrderTypes + limit orders)
- CLAUDE.md: add GitHub issues/SPEC workflow section

+++ freqpred/strategy/config.py
@@ -13,8 +13,8 @@ class StrategyConfig:
     kelly_fraction: float
     categories: list[str]
     min_volume_24h: float
-    max_days_to_close: int
-    min_days_to_close: int
+    max_days_to_close: float
+    min_days_to_close: float
 
     # Exit management
     stoploss: float = -0.20
